Rename -> location_type to cold_storage_type
