+++ server/service/src/sync/test/test_data/cold_storage_type.rs
