+++ server/service/src/sync/translations/mod.rs
@@ -10,6 +10,7 @@ pub(crate) mod asset_type;
 pub(crate) mod barcode;
 pub(crate) mod clinician;
 pub(crate) mod clinician_store_join;
+pub(crate) mod cold_storage_type;
 pub(crate) mod currency;
 pub(crate) mod demographic;
 pub(crate) mod document;
@@ -21,7 +22,6 @@ pub(crate) mod item;
 pub(crate) mod item_variant;
 pub(crate) mod location;
 pub(crate) mod location_movement;
-pub(crate) mod location_type;
 pub(crate) mod master_list;
 pub(crate) mod master_list_line;
 pub(crate) mod master_list_name_join;
@@ -94,7 +94,7 @@ pub(crate) fn all_translators() -> SyncTranslators {
         document_registry::boxed(),
         property::boxed(),
         name_property::boxed(),
-        location_type::boxed(), // Note: Translates to cold_storage_type
+        cold_storage_type::boxed(),
         // Remote
         location::boxed(),
         location_movement::boxed(),
